% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview_chrome.R
\name{preview_chrome}
\alias{preview_chrome}
\title{Preview cheat sheet in chrome browser}
\usage{
preview_chrome(input, chrome = TRUE)
}
\arguments{
\item{input}{(`character(1)`)\cr
The path to the input R Markdown document (.Rmd) to be previewed.}

\item{chrome}{(`logical(1)`)\cr
Use chrome or chromium?}
}
\description{
Preview cheat sheets in chrome or chromium browser.
}
