% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_pdf.R
\name{print_pdf}
\alias{print_pdf}
\title{Print cheat sheet to pdf}
\usage{
print_pdf(input, output)
}
\arguments{
\item{input}{(`character(1)`)\cr
The path to the input R Markdown document (.Rmd) to be rendered.}

\item{output}{(`character(1)`)\cr
The path of the output file (.pdf).}
}
\description{
The `print_pdf()` function renders the input R Markdown file and prints the
page in a chrome headless session.
}
